@@ -27,87 +24,94 @@
   },
   "engineStrict": true,
   "engines": {
-    "node": ">=6.9.0"
+    "node": ">=7.0.0"
   },
   "dependencies": {
-    "amd-loader": "0.0.5",
-    "autoprefixer": "^6.7.6",
+    "amd-loader": "0.0.8",
+    "autoprefixer": "^7.1.2",
     "babel-cli": "^6.7.5",
-    "babel-core": "^6.17.0",
+    "babel-core": "^6.25.0",
     "babel-eslint": "^7.1.0",
-    "babel-loader": "^6.2.4",
+    "babel-loader": "^7.1.1",
     "babel-plugin-module-resolver": "^2.2.0",
     "babel-plugin-transform-async-to-generator": "^6.16.0",
-    "babel-plugin-transform-es2015-block-scoping": "^6.7.1",
-    "babel-plugin-transform-es2015-destructuring": "^6.6.5",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.22.0",
-    "babel-plugin-transform-es2015-parameters": "^6.7.0",
-    "babel-plugin-transform-es2015-spread": "^6.6.5",
     "babel-plugin-transform-flow-strip-types": "^6.14.0",
     "babel-plugin-transform-runtime": "^6.7.5",
     "babel-plugin-webpack-alias": "^2.1.1",
     "babel-polyfill": "^6.7.4",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babel-register": "^6.18.0",
     "body-parser": "^1.15.2",
-    "check-node-version": "^1.1.2",
+    "check-node-version": "^3.2.0",
     "chrome-remote-interface": "0.17.0",
     "classnames": "^2.2.5",
     "co": "=4.6.0",
     "css-loader": "^0.26.1",
-    "devtools-config": "^0.0.15",
-    "devtools-connection": "^0.0.5",
-    "devtools-modules": "^0.0.28",
+    "debug": "^3.1.0",
+    "devtools-config": "^0.0.16",
+    "devtools-connection": "^1.0.7",
+    "devtools-contextmenu": "~1.0.10",
+    "devtools-environment": "^0.0.5",
+    "devtools-license-check": "^0.7.0",
+    "devtools-mc-assets": "^0.0.7",
+    "devtools-modules": "~1.1.6",
     "devtools-sprintf-js": "^1.0.3",
-    "devtools-utils": "^0.0.2",
-    "eslint": "^3.12.0",
-    "eslint-plugin-babel": "^3.3.0",
-    "eslint-plugin-flowtype": "^2.20.0",
-    "eslint-plugin-mozilla": "0.2.3",
-    "eslint-plugin-react": "^6.7.1",
     "express": "^4.13.4",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "express-static": "^1.2.5",
+    "extract-text-webpack-plugin": "^3.0.0",
     "fs-extra": "^2.0.0",
     "fuzzaldrin-plus": "^0.4.0",
-    "geckodriver": "=1.4.0",
+    "geckodriver": "=1.12.2",
     "immutable": "^3.7.6",
     "json-loader": "^0.5.4",
+    "md5": "^2.2.1",
+    "mime-types": "^2.1.18",
     "minimist": "^1.2.0",
     "mustache": "^2.2.1",
     "node-static": "^0.7.7",
-    "postcss": "^5.2.12",
-    "postcss-bidirection": "=2.3.0",
-    "postcss-loader": "^1.2.2",
+    "postcss": "^6.0.6",
+    "postcss-class-namespace": "^0.1.0",
+    "postcss-loader": "^2.0.6",
+    "postcss-url-mapper": "^1.2.0",
+    "prop-types": "^15.6.0",
     "properties-parser": "^0.3.1",
     "ps-node": "^0.1.4",
     "raw-loader": "^0.5.1",
-    "react": "=15.3.2",
-    "react-dom": "=15.3.2",
-    "react-hot-loader": "^1.3.1",
-    "react-immutable-proptypes": "^2.1.0",
-    "react-inlinesvg": "^0.5.3",
-    "react-redux": "4.4.5",
-    "redux": "3.5.2",
-    "selenium-webdriver": "=3.3.0",
-    "style-loader": "^0.13.1",
-    "svg-inline-loader": "^0.7.1",
-    "webpack": "1.14.0",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-env-loader-plugin": "^0.1.4",
-    "webpack-hot-middleware": "^2.12.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-dom-factories": "^1.0.2",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
+    "selenium-webdriver": "=3.6.0",
+    "style-loader": "^0.18.2",
+    "svg-inline-loader": "^0.8.0",
+    "svg-inline-react": "3.0.0",
+    "url-loader": "^0.5.9",
+    "webpack": "^3.3.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-env-loader-plugin": "^1.0.0",
+    "webpack-hot-middleware": "^2.18.2",
     "ws": "^1.0.1"
   },
   "devDependencies": {
-    "@kadira/storybook": "^2.35.3",
+    "eslint": "^3.12.0",
+    "eslint-plugin-babel": "^3.3.0",
+    "eslint-plugin-flowtype": "^2.20.0",
+    "eslint-plugin-mozilla": "0.4.3",
+    "eslint-plugin-react": "^6.7.1",
     "ipaddr": "^0.0.9",
-    "yarn": "^0.20.3",
-    "jest": "^19.0.2"
+    "jest": "^20.0.4",
+    "react-hot-loader": "^1.3.1"
   },
   "files": [
+    "assets",
     "bin",
     "src",
     "configs",
     "index.html",
     "index.js",
+    "postcss.config.js",
     "webpack.config.js",
     "webpack.config.devtools.js"
   ],
