@@ -1,22 +1,22 @@
 {
   "name": "devtools-splitter",
-  "version": "0.0.3",
+  "version": "0.0.7",
   "description": "DevTools Splitter",
   "main": "index.js",
   "scripts": {
     "storybook": "start-storybook -p 9001 -c .storybook",
-    "snapshot": "build-storybook && percy-storybook",
     "license-check": "devtools-license-check",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "# No tests available"
   },
   "author": "Jason Laster",
   "license": "MPL-2.0",
   "devDependencies": {
     "@kadira/storybook": "^2.35.3",
-    "react": "=15.3.2",
-    "react-dom": "=15.3.2"
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2"
   },
   "dependencies": {
-    "@percy-io/react-percy-storybook": "^0.1.2"
+    "prop-types": "^15.6.0",
+    "react-dom-factories": "^1.0.2"
   }
 }
