@@ -4,18 +4,27 @@
 
 const React = require("react");
 const ReactDOM = require("react-dom");
-const { DOM: dom, PropTypes } = React;
+const { Component } = React;
+const PropTypes = require("prop-types");
+const dom = require("react-dom-factories");
 
-const Draggable = React.createClass({
-  displayName: "Draggable",
+class Draggable extends Component {
+  static get propTypes() {
+    return {
+      onMove: PropTypes.func.isRequired,
+      onStart: PropTypes.func,
+      onStop: PropTypes.func,
+      style: PropTypes.object,
+      className: PropTypes.string
+    };
+  }
 
-  propTypes: {
-    onMove: PropTypes.func.isRequired,
-    onStart: PropTypes.func,
-    onStop: PropTypes.func,
-    style: PropTypes.object,
-    className: PropTypes.string
-  },
+  constructor(props) {
+    super(props);
+    this.startDragging = this.startDragging.bind(this);
+    this.onMove = this.onMove.bind(this);
+    this.onUp = this.onUp.bind(this);
+  }
 
   startDragging(ev) {
     ev.preventDefault();
@@ -23,14 +32,14 @@ const Draggable = React.createClass({
     doc.addEventListener("mousemove", this.onMove);
     doc.addEventListener("mouseup", this.onUp);
     this.props.onStart && this.props.onStart();
-  },
+  }
 
   onMove(ev) {
     ev.preventDefault();
     // We pass the whole event because we don't know which properties
     // the callee needs.
     this.props.onMove(ev);
-  },
+  }
 
   onUp(ev) {
     ev.preventDefault();
@@ -38,7 +47,7 @@ const Draggable = React.createClass({
     doc.removeEventListener("mousemove", this.onMove);
     doc.removeEventListener("mouseup", this.onUp);
     this.props.onStop && this.props.onStop();
-  },
+  }
 
   render() {
     return dom.div({
@@ -47,6 +56,6 @@ const Draggable = React.createClass({
       onMouseDown: this.startDragging
     });
   }
-});
+}
 
 module.exports = Draggable;
