@@ -2,9 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const fs = require('fs');
+const fs = require("fs");
 const path = require("path");
-const parse = require("yarn/lib/lockfile/parse.js").default;
 
 function isUnique(list) {
   const set = new Set(list);
@@ -22,6 +21,10 @@ const depsList = devtoolDeps.map(dep => dep.replace(/@.*/, ""));
 const hasDuplicates = !isUnique(depsList);
 
 if (hasDuplicates) {
-  console.warn(`Oops, it looks like there's duplicate dependencies:\n\n${devtoolDeps.join("\n")}\n\nCheck your yarn.lock file for the discrepancy.`);
+  console.warn(
+    `Oops, it looks like there's duplicate dependencies:\n\n${devtoolDeps.join(
+      "\n"
+    )}\n\nCheck your yarn.lock file for the discrepancy.`
+  );
   process.exit(1);
 }
