@@ -7,7 +7,7 @@ const {
   DebuggerClient,
   DebuggerTransport,
   TargetFactory,
-  WebsocketTransport,
+  WebsocketTransport
 } = require("devtools-connection");
 
 const { getValue } = require("devtools-config");
@@ -17,7 +17,7 @@ import type {
   TabTarget,
   TabPayload,
   DebuggerClient as DebuggerClientType,
-  ThreadClient,
+  ThreadClient
 } from "./firefox-types";
 
 let debuggerClient: DebuggerClientType | null = null;
@@ -34,21 +34,20 @@ function createTabs(tabs: TabPayload[]): Tab[] {
       url: tab.url,
       id: tab.actor,
       tab,
-      clientType: "firefox",
+      clientType: "firefox"
     };
   });
 }
 
 async function connectClient() {
-  const useProxy = !getValue("firefox.webSocketConnection");
-  const firefoxHost = getValue(
-    useProxy ? "firefox.proxyHost" : "firefox.webSocketHost",
-  );
+  const useWebSocket = getValue("firefox.webSocketConnection");
+  const firefoxHost = useWebSocket ? getValue("firefox.host") : "localhost";
+  const firefoxPort = getValue("firefox.webSocketPort");
 
-  const socket = new WebSocket(`ws://${firefoxHost}`);
-  const transport = useProxy
-    ? new DebuggerTransport(socket)
-    : new WebsocketTransport(socket);
+  const socket = new WebSocket(`ws://${firefoxHost}:${firefoxPort}`);
+  const transport = useWebSocket
+    ? new WebsocketTransport(socket)
+    : new DebuggerTransport(socket);
 
   debuggerClient = new DebuggerClient(transport);
   if (!debuggerClient) {
@@ -57,8 +56,7 @@ async function connectClient() {
 
   try {
     await debuggerClient.connect();
-    const tabs = await getTabs();
-    return tabs;
+    return await getTabs();
   } catch (err) {
     console.log(err);
     return [];
@@ -74,11 +72,9 @@ async function connectTab(tab: Tab) {
 
   const tabTarget: TabTarget = await lookupTabTarget(tab);
 
-  const [, threadClient: ThreadClient] = await tabTarget.activeTab.attachThread(
-    {
-      ignoreFrameEnvironment: true
-    },
-  );
+  const [, threadClient: ThreadClient] = await tabTarget.activeTab.attachThread({
+    ignoreFrameEnvironment: true
+  });
 
   threadClient.resume();
   return { debuggerClient, threadClient, tabTarget };
@@ -93,11 +89,11 @@ async function getTabs() {
   return createTabs(response.tabs);
 }
 
-function initPage() {}
+function initPage(options: Object) {}
 
 module.exports = {
   connectClient,
   connectTab,
   initPage,
-  getTabs,
+  getTabs
 };
