@@ -6,17 +6,24 @@ const ps = require("child_process");
 const path = require("path");
 const { isFirefoxRunning } = require("./utils/firefox");
 const firefoxDriver = require("../../bin/firefox-driver");
+const {
+  getValue
+} = require("devtools-config");
 
 function handleLaunchRequest(req, res) {
   const browser = req.body.browser;
-  const location = "https://devtools-html.github.io/debugger-examples/";
+  const location = getValue("defaultURL");
 
   process.env.PATH += `:${__dirname}`;
   if (browser == "Firefox") {
     isFirefoxRunning().then((isRunning) => {
-      console.log("running", isRunning);
+      const options = {
+        useWebSocket: getValue("firefox.webSocketConnection"),
+        webSocketPort: getValue("firefox.webSocketPort"),
+        tcpPort: getValue("firefox.tcpPort")
+      };
       if (!isRunning) {
-        firefoxDriver.start(location);
+        firefoxDriver.start(location, options);
         res.end("launched firefox");
       } else {
         res.end("already running firefox");
@@ -25,8 +32,8 @@ function handleLaunchRequest(req, res) {
   }
 
   if (browser == "Chrome") {
-    ps.spawn(path.resolve(__dirname, "../../bin/chrome-driver.js"),
-     ["--location", location]);
+    const chromeDriver = path.resolve(__dirname, "../../bin/chrome-driver.js");
+    ps.spawn("node", [chromeDriver, "--location", location]);
     res.end("launched chrome");
   }
 }
