@@ -3,10 +3,12 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const React = require("react");
-const { storiesOf, action } = require("@kadira/storybook");
+const { storiesOf, action } = require("@storybook/react");
 const LandingPage = require("../src/components/LandingPage");
 
 const { combineReducers } = require("redux");
+const dom = require("react-dom-factories");
+
 let reducers = require("../src/reducers");
 let constants = require("../src/constants");
 let getState = combineReducers(reducers);
@@ -15,7 +17,7 @@ const selectors = require("../src/selectors");
 import { Map } from "immutable";
 
 // Add devtools theme styles
-require("../src/lib/themes/light-theme.css");
+// require("../src/lib/themes/light-theme.css");
 document.body.classList.add("theme-light");
 
 const getTab = (id, title, clientType, url) => {
@@ -41,7 +43,7 @@ const getTabs = (tabs, state) => {
 };
 
 const renderLandingPage = (props) => {
-  return React.DOM.div({}, React.createElement(LandingPage, Object.assign({
+  return dom.div({}, React.createElement(LandingPage, Object.assign({
     onFilterChange: action("onFilterChange"),
     onTabClick: action("onTabClick"),
     title: "Storybook test",
