@@ -1,28 +1,32 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 require("babel-register");
 
 const path = require("path");
+
 const webpack = require("webpack");
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
-const { isDevelopment, isFirefoxPanel, getValue } = require("devtools-config");
+const {isDevelopment, isFirefoxPanel} = require("devtools-environment")
+const { getValue, setConfig } = require("devtools-config");
 const NODE_ENV = process.env.NODE_ENV || "development";
 const TARGET = process.env.TARGET || "local";
 
-const defaultBabelPlugins = [
-  "transform-flow-strip-types",
-  "transform-async-to-generator"
-];
+module.exports = (webpackConfig, envConfig, options = {}) => {
+  setConfig(envConfig);
 
-module.exports = (webpackConfig, envConfig) => {
   webpackConfig.context = path.resolve(__dirname, "src");
   webpackConfig.devtool = "source-map";
 
   webpackConfig.module = webpackConfig.module || {};
-  webpackConfig.module.loaders = webpackConfig.module.loaders || [];
-  webpackConfig.module.loaders.push({
+  webpackConfig.module.rules = webpackConfig.module.rules || [];
+  webpackConfig.module.rules.push({
     test: /\.json$/,
-    loader: "json"
+    loader: "json-loader"
   });
-  webpackConfig.module.loaders.push({
+
+  webpackConfig.module.rules.push({
     test: /\.js$/,
     exclude: request => {
       // Some paths are excluded from Babel
@@ -30,41 +34,26 @@ module.exports = (webpackConfig, envConfig) => {
       let excludedRe = new RegExp(`(${excludedPaths.join("|")})`);
       let excluded = !!request.match(excludedRe);
 
-      if (webpackConfig.babelExcludes) {
+      if (options.babelExcludes) {
         // If the tool defines an additional exclude regexp for Babel.
-        excluded = excluded || !!request.match(webpackConfig.babelExcludes);
+        excluded = excluded || !!request.match(options.babelExcludes);
       }
 
-      return excluded && !request.match(/devtools-/);
+      let included = ["devtools-"]
+      if (options.babelIncludes) {
+        included = included.concat(options.babelIncludes);
+      }
+
+      const reincludeRe = new RegExp(`node_modules(\\/|\\\\)${included.join("|")}`);
+      return excluded && !request.match(reincludeRe);
     },
-    loaders: [
-      `babel?${defaultBabelPlugins.map(p => `plugins[]=${p}`)}&ignore=src/lib`
-    ],
-    isJavaScriptLoader: true
-  });
-  webpackConfig.module.loaders.push({
-    test: /\.svg$/,
-    loader: "svg-inline"
+    loader: `babel-loader?ignore=src/lib`
   });
 
-  webpackConfig.module.loaders.push({
+  webpackConfig.module.rules.push({
     test: /\.properties$/,
-    loader: "raw"
+    loader: "raw-loader"
   });
-
-  // Add resolveLoader for ./node_modules to fix issues when synlinked.
-  webpackConfig.resolveLoader = webpackConfig.resolveLoader || {};
-  webpackConfig.resolveLoader.root = webpackConfig.resolveLoader.root || [];
-  webpackConfig.resolveLoader.root.push(path.resolve("./node_modules"));
-
-  // Add a resolve alias for React if the target is UMD
-  if (webpackConfig.output && webpackConfig.output.libraryTarget === "umd") {
-    webpackConfig.resolve = webpackConfig.resolve || {};
-    webpackConfig.resolve.alias = webpackConfig.resolve.alias || {};
-    if (!webpackConfig.resolve.alias.react) {
-      webpackConfig.resolve.alias.react = "react/lib/ReactUMDEntry";
-    }
-  }
 
   webpackConfig.node = { fs: "empty" };
 
@@ -80,30 +69,96 @@ module.exports = (webpackConfig, envConfig) => {
   );
 
   if (isDevelopment()) {
-    webpackConfig.module.loaders.push({
-      test: /\.css$/,
-      loader: "style!css!postcss"
+    /*
+     * SVGs are loaded in one of two ways in JS w/ SVG inline loader
+     * and in CSS w/ the CSS loader.
+     *
+     * Inline SVGs are included in the JS bundle and mounted w/ React.
+     *
+     * SVG URLs like chrome://devtools/skin/images/arrow.svg are mapped
+     * by the postcss-loader to /mc/devtools/client/themes/arrow.svg
+     * and are hosted in `development-server` with an express server.
+     *
+     * CSS URLs like resource://devtools/client/themes/variables.css are mapped by the
+     * NormalModuleReplacementPlugin to
+     * devtools-mc-assets/assets/devtools/client/themes/arrow.svg.
+     * The modules are then resolved by the css-loader, which allows modules like
+     * variables.css to be bundled.
+     *
+     * We use several PostCSS plugins to make local development a little easier:
+     * autoprefixer, bidirection. These plugins help support chrome + firefox
+     * development w/ new CSS features like RTL, mask, ...
+     */
+
+    webpackConfig.module.rules.push({
+      test: /svg$/,
+      loader: "svg-inline-loader"
     });
 
-    if (getValue("hotReloading")) {
-      Object.keys(webpackConfig.entry).forEach(key => {
-        webpackConfig.entry[key].push("webpack-hot-middleware/client");
-      });
+    const cssUses = [{
+      loader: "style-loader"
+    }, {
+      loader: "css-loader",
+      options: {
+        importLoaders: 1,
+        url: false
+      }
+    }];
 
-      webpackConfig.plugins = webpackConfig.plugins.concat([
-        new webpack.HotModuleReplacementPlugin(),
-        new webpack.NoErrorsPlugin()
-      ]);
-
-      webpackConfig.module.loaders.forEach(spec => {
-        if (spec.isJavaScriptLoader) {
-          spec.loaders.unshift("react-hot");
-        }
-      });
+    if (!options.disablePostCSS) {
+      cssUses.push({ loader: "postcss-loader" });
     }
+
+    webpackConfig.module.rules.push({
+      test: /\.css$/,
+      use: cssUses
+    });
+
+    webpackConfig.plugins.push(
+      new webpack.NormalModuleReplacementPlugin(
+        /(resource:|chrome:).*.css/,
+        function (resource) {
+          const newUrl = resource.request
+            .replace(
+              /(.\/chrome:\/\/|.\/resource:\/\/)/,
+              `devtools-mc-assets/assets/`
+            )
+            .replace(/devtools\/skin/, "devtools/client/themes")
+            .replace(/devtools\/content/, "devtools/client");
+
+          resource.request = newUrl;
+        }
+      )
+    );
   } else {
+    /*
+    * SVGs are loaded in one of two ways in JS w/ SVG inline loader
+    * and in CSS w/ the CSS loader.
+    *
+    * Inline SVGs are included in the JS bundle and mounted w/ React.
+    *
+    * SVG URLs like /images/arrow.svg are mapped
+    * by the postcss-loader to chrome://chrome://devtools/skin/images/debugger/arrow.svg,
+    * copied to devtools/themes/images/debugger and added to the jar.mn
+    *
+    * SVG URLS like chrome://devtools/skin/images/add.svg are
+    * ignored as they are already available in MC.
+    */
+
+    const cssUses = [{
+      loader: "css-loader",
+      options: {
+        importLoaders: 1,
+        url: false
+      }
+    }];
+
+    if (!options.disablePostCSS) {
+      cssUses.push({ loader: "postcss-loader" });
+    }
+
     // Extract CSS into a single file
-    webpackConfig.module.loaders.push({
+    webpackConfig.module.rules.push({
       test: /\.css$/,
       exclude: request => {
         // If the tool defines an exclude regexp for CSS files.
@@ -111,35 +166,37 @@ module.exports = (webpackConfig, envConfig) => {
           webpackConfig.cssExcludes && request.match(webpackConfig.cssExcludes)
         );
       },
-      loader: ExtractTextPlugin.extract(
-        "style-loader",
-        "css-loader",
-        "postcss-loader"
-      )
+      use: ExtractTextPlugin.extract({
+        filename: "*.css",
+        use: cssUses
+      })
+    });
+
+    webpackConfig.module.rules.push({
+      test: /svg$/,
+      loader: "svg-inline-loader"
     });
 
     webpackConfig.plugins.push(new ExtractTextPlugin("[name].css"));
   }
 
-  webpackConfig.postcss = () => [
-    require("postcss-bidirection"),
-    require("autoprefixer")
-  ];
-
   if (isFirefoxPanel()) {
     webpackConfig = require("./webpack.config.devtools")(
       webpackConfig,
-      envConfig
+      envConfig,
+      options
     );
   }
 
   // NOTE: This is only needed to fix a bug with chrome devtools' debugger and
-  // destructuring params https://github.com/devtools-html/debugger.html/issues/67
+  // destructuring params https://github.com/firefox-devtools/debugger.html/issues/67
   if (getValue("transformParameters")) {
-    webpackConfig.module.loaders.forEach(spec => {
+    webpackConfig.module.rules.forEach(spec => {
       if (spec.isJavaScriptLoader) {
-        const idx = spec.loaders.findIndex(loader => loader.includes("babel"));
-        spec.loaders[idx] += "&plugins[]=transform-es2015-parameters";
+        const idx = spec.rules.findIndex(loader =>
+          loader.includes("babel-loader")
+        );
+        spec.rules[idx] += "&plugins[]=transform-es2015-parameters";
       }
     });
   }
