@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const { appinfo } = require("./Services");
+const { appinfo } = require("devtools-services");
 const EventEmitter = require("./utils/event-emitter");
 const isOSX = appinfo.OS === "Darwin";
 
@@ -240,4 +240,4 @@ KeyShortcuts.prototype = {
     this.eventEmitter.off(key, listener);
   },
 };
-exports.KeyShortcuts = KeyShortcuts;
+module.exports = KeyShortcuts;
